[KERNELS] Consolidate Blackwell FP4 layout metadata and tests (#11576)
Blackwell FP4 conversion had a separate column-major stride algorithm,
and the shuffled-layout converter imported it to allocate plain-layout
output. Make physical strides part of
`BlackwellMXValueLayoutTransformation` and derive them through the
existing `StridedLayout` implementation. All three allocation sites now
use layout-owned metadata, removing the duplicate algorithm and keeping
the physical layout definition in one place.

The two direct-conversion calls use `major_dim=-2`, which the shared
converter already normalizes. GPU kernels, packing, padding semantics,
and output allocation sizes are unchanged. This follows [Convert between
Blackwell FP4 layouts
directly](https://github.com/triton-lang/triton/pull/11563).

Move the Blackwell-specific conversion and allocation tests beside the
existing Blackwell tests. Keep all 96 CUDA conversion cases, including
empty inputs and supplied output buffers, while reducing repeated
CPU/meta fallback cases from 192 to 32. Check output device, default
plain-layout strides, and shuffled contiguity; value equality alone
would miss an incorrectly ordered allocation. Extend the existing
empty-shape roundtrip cases with plain → shuffled → plain, covering
empty K/N and leading dimensions. Padding poisoning, output sentinels,
and all four allocation cases remain covered.

## Validation

Validated baseline and branch package snapshots on the same NVIDIA GB300
with the installed Triton 3.8.0 compiler, PyTorch 2.11.0, and CUDA 13.0.
The baseline passed 837 selected cases; this branch passed 719, with no
skips in either run. The reduction is 160 repeated fallback cases
removed and 42 empty-shape cases added.

The selection includes the complete Blackwell layout test module plus
generic FP4 output-buffer, meta-device, and storage-shape tests. It
exercises both conversion directions, ranks 2–5, irregular tiles, sliced
storage, padding, output layout, and all four direct-conversion
allocation cases. Remote source hashes match the committed files. This
validates the Python package changes against the installed compiler;
Triton main's compiler was not rebuilt.

`pre-commit run --from-ref origin/main --to-ref HEAD` passed.

## Lines changed

| Files | Added | Removed | Net |
| --- | ---: | ---: | ---: |
| Tests | +92 | -79 | +13 |
| Non-test | +11 | -21 | -10 |
| All | +103 | -100 | +3 |

## Reviewer's guide

- `python/triton_kernels/triton_kernels/tensor_details/layout_details/`:
consolidate plain Blackwell storage metadata through the existing
strided implementation and simplify direct-conversion calls.
- `python/triton_kernels/tests/`: move the two Blackwell-specific tests
from `test_tensor.py` to `test_tensor_details/test_layout_blackwell.py`,
reduce fallback repetition, and strengthen output-layout and empty-shape
coverage.

<!-- codex-thread: 01a064c1-5fae-7e31-9f79-2aeb8e10070d -->
