+++ python/triton_kernels/tests/test_tensor_details/test_layout_blackwell.py
@@ -46,6 +46,7 @@ def test_scale_zero_sized_roundtrip(shape, layout, device):
 @pytest.mark.parametrize("layouts", [
     (BlackwellMXValueLayout(), ),
     (BlackwellMX4ValueShuffledLayout(), BlackwellMX4ValueShuffledLayout(block_k=256, block_n=128)),
+    (BlackwellMXValueLayout(), BlackwellMX4ValueShuffledLayout(), BlackwellMXValueLayout()),
 ])
 @pytest.mark.parametrize("device", ["cpu", "meta", "cuda"])
 def test_value_zero_sized_roundtrip(shape, major_dim, layouts, device):
@@ -151,6 +152,97 @@ def test_mxfp4_value_convert_layout_matches_torch(shape, layout, trans, step):
         assert actual.storage.data.is_contiguous()
         assert torch.equal(actual.storage.data.cpu(), expected.storage.data)
     assert torch.equal(roundtrip.storage.data, data_cuda)
+
+
+@pytest.mark.parametrize(("shape", "device"), [
+    ((256, 512), "cuda"),
+    ((130, 66), "cuda"),
+    ((2, 130, 66), "cuda"),
+    ((2, 3, 130, 66), "cuda"),
+    ((2, 1, 3, 130, 66), "cuda"),
+    ((0, 130, 66), "cuda"),
+    ((2, 130, 66), "cpu"),
+    ((2, 130, 66), "meta"),
+])
+@pytest.mark.parametrize("tiles", [(128, 256), (192, 48)])
+@pytest.mark.parametrize("step", [1, 2])
+@pytest.mark.parametrize("with_out", [False, True])
+@pytest.mark.parametrize("inverse", [False, True])
+def test_blackwell_fp4_shuffle_conversion(shape, tiles, step, with_out, device, inverse):
+    data = torch.randint(0, 256, (*shape[:-1], shape[-1] // 2), dtype=torch.uint8,
+                         generator=torch.Generator().manual_seed(0))
+    canonical = wrap_torch_tensor(data, dtype=FP4)
+    plain, shuffled = BlackwellMXValueLayout(), BlackwellMX4ValueShuffledLayout(*tiles)
+    source_layout, destination = (shuffled, plain) if inverse else (plain, shuffled)
+    source = convert_layout(canonical, source_layout)
+    expected = convert_layout(canonical, destination)
+    # Padding bytes must not leak into logical values or shuffled destination padding.
+    if inverse:
+        _, tiles_k, tiles_n, tile_n, tile_k = source.data.shape
+        k = torch.arange(tiles_k * tile_k).reshape(1, tiles_k, 1, 1, tile_k)
+        n = torch.arange(tiles_n * tile_n).reshape(1, 1, tiles_n, tile_n, 1)
+        source.data.masked_fill_((k >= shape[-2] // 2) | (n >= shape[-1]), 0xAB)
+    else:
+        source.data[..., shape[-2] // 2:, :].fill_(0xAB)
+    source_data = source.data.to(device)
+    if step == 2:
+        storage = torch.empty(tuple(2 * size for size in source_data.shape), dtype=torch.uint8, device=device)
+        source_view = storage[tuple(slice(None, None, 2) for _ in range(source_data.ndim))]
+        source_view.copy_(source_data)
+        source_data = source_view
+    source = wrap_torch_tensor(source_data, dtype=FP4, shape=shape, layout=source.storage.layout)
+    out = None
+    if with_out:
+        storage = torch.full((expected.data.numel() * step + 1, ), 0xCD, dtype=torch.uint8, device=device)
+        out_data = storage[1:].as_strided(expected.data.shape, tuple(step * s for s in expected.data.stride()))
+        out = wrap_torch_tensor(out_data, dtype=FP4, shape=shape, layout=destination)
+
+    actual = convert_layout(source, destination, out=out)
+
+    assert actual.shape == list(shape)
+    assert actual.data.shape == expected.data.shape
+    assert actual.data.dtype == expected.data.dtype
+    assert actual.device == source.device
+    if with_out:
+        assert actual is out
+    elif inverse:
+        assert actual.data.stride() == expected.data.stride()
+    else:
+        assert actual.data.is_contiguous()
+    if device != "meta":
+        actual_data, expected_data = actual.data.cpu(), expected.data
+        if inverse:
+            actual_data = actual_data[..., :shape[-2] // 2, :]
+            expected_data = expected_data[..., :shape[-2] // 2, :]
+            if device == "cuda" and with_out:
+                assert torch.all(out.data[..., shape[-2] // 2:, :] == 0xCD)
+        assert torch.equal(actual_data, expected_data)
+        assert torch.equal(convert_layout(actual, StridedLayout()).data.cpu(), data)
+        if with_out:
+            assert storage[0].item() == 0xCD
+            if step == 2:
+                assert torch.all(storage[2::2] == 0xCD)
+
+
+@pytest.mark.parametrize("with_out", [False, True])
+@pytest.mark.parametrize("inverse", [False, True])
+def test_blackwell_fp4_shuffle_peak_allocation(with_out, inverse):
+    canonical = empty((8, 1024, 2048), dtype=FP4, device="cuda")
+    plain, shuffled = BlackwellMXValueLayout(), BlackwellMX4ValueShuffledLayout()
+    source_layout, destination = (shuffled, plain) if inverse else (plain, shuffled)
+    source = convert_layout(canonical, source_layout)
+    warm = convert_layout(source, destination)
+    out = warm if with_out else None
+    torch.cuda.synchronize()
+    del warm
+    baseline = torch.cuda.memory_allocated()
+    torch.cuda.reset_peak_memory_stats()
+
+    actual = convert_layout(source, destination, out=out)
+    torch.cuda.synchronize()
+    peak = torch.cuda.max_memory_allocated() - baseline
+
+    assert peak <= (0 if with_out else actual.data.nbytes) + 1024**2
 
 
 @pytest.mark.parametrize("shape", [(256, 128), (2, 258, 257), (2, 3, 130, 65)])
