+++ python/triton_kernels/triton_kernels/tensor_details/layout_details/blackwell_value.py
@@ -20,18 +20,6 @@ class BlackwellMXValueLayout(Layout):
         return block_shape
 
 
-def strides_major_dim_m2(shape):
-    n = len(shape)
-    if n <= 1:
-        return [1] * n
-    order = [n - 2, n - 1] + list(range(n - 3, -1, -1))  # fastest -> slowest
-    st = [0] * n
-    st[order[0]] = 1
-    for prev, d in zip(order, order[1:]):
-        st[d] = st[prev] * shape[prev]
-    return st
-
-
 # ------------------- Blackwell MX Value Layout Transformation -------------------
 @dataclass(frozen=True)
 class BlackwellMXValueLayoutTransformation(LayoutTransformation):
@@ -45,6 +33,11 @@ class BlackwellMXValueLayoutTransformation(LayoutTransformation):
         M //= 2
         M += -M % 128
         return [*leading_shape, M, K]
+
+    @property
+    def storage_strides(self) -> list[int]:
+        # The physical shape already includes packing and TMA padding.
+        return strided.StridedLayout(-2).make_transformation(self.storage_shape, is_fp4=False).storage_strides
 
     def convert_data(self, data, destination: LayoutTransformation, *, out=None):
         if (not self.is_fp4 or data.device.type != "cuda" or data.dtype != torch.uint8 or self.shape[-2] % 2
@@ -64,15 +57,13 @@ class BlackwellMXValueLayoutTransformation(LayoutTransformation):
                 or self.shape[-2] % 2 or not source._can_convert_fp4(data)):
             return super()._convert_data_from(data, source, out=out)
         if out is None:
-            out = torch.empty_strided(self.storage_shape, strides_major_dim_m2(self.storage_shape), device=data.device,
-                                      dtype=data.dtype)
+            out = torch.empty_strided(self.storage_shape, self.storage_strides, device=data.device, dtype=data.dtype)
         repack(data, source.order[0], -2, True, out=out[..., :self.shape[-2] // 2, :])
         return out
 
     def swizzle_data(self, data):
         # re-pack as column-major
-        ret = torch.empty_strided(self.storage_shape, strides_major_dim_m2(self.storage_shape), device=data.device,
-                                  dtype=data.dtype)
+        ret = torch.empty_strided(self.storage_shape, self.storage_strides, device=data.device, dtype=data.dtype)
         repacked_shape = list(data.shape)
         repacked_shape[-1] *= 2
         repacked_shape[-2] //= 2
